Revert "Js testing"
