+++ .eslintrc.json
@@ -1,8 +1,7 @@
 {
     "env": {
         "browser": true,
-        "es2021": true,
-        "jest/globals": true
+        "es2021": true
     },
     "extends": [
         "plugin:react/recommended",
@@ -17,8 +16,7 @@
     },
     "plugins": [
         "react",
-        "unused-imports",
-        "jest"
+        "unused-imports"
     ],
     "rules": {
         "react/react-in-jsx-scope": "off",
