+++ components/BaseMap.js
@@ -1,3 +1,10 @@
+export function initGoogleMap (googleMapRef, center, zoom) {
+  return new window.google.maps.Map(googleMapRef.current, {
+    center,
+    zoom
+  })
+}
+
 export function addMarker (location, map, draggable) {
   // Add the marker at the clicked location, and add the next-available label
   // from the array of alphabetical characters.
