+++ components/DataGridTable.js
@@ -5,12 +5,12 @@ import { DataGrid, esES, GridActionsCellItem, GridToolbar } from '@mui/x-data-gr
 import moment from 'moment'
 import Link from 'next/link'
 import { useCallback, useContext, useState } from 'react'
-import { conf } from '../../configuration'
-import { PdrContext } from '../../context/PdrContext'
-import { TownContext } from '../../context/TownContext'
-import { calculateAlert } from '../../utils/pdr-management'
-import DeleteRowDialog from '../DeleteRowDialog'
-import { GreenRadio, RedRadio, YellowRadio } from '../RadioButtons'
+import { conf } from '../configuration'
+import { PdrContext } from '../context/PdrContext'
+import { TownContext } from '../context/TownContext'
+import { calculateAlert } from '../utils/pdr-management'
+import DeleteRowDialog from './DeleteRowDialog'
+import { GreenRadio, RedRadio, YellowRadio } from './RadioButtons'
 
 export function lastNweeks (recogida, n) {
   const fiveMondays = [...Array(5).keys()].map(nWeek => moment().day('monday').subtract(nWeek, 'weeks'))
@@ -188,7 +188,7 @@ export default function DataGridTable () {
   }
 
   return (
-    <div style={{ height: '80vh', width: '100%' }} data-testid="table">
+    <div style={{ height: '80vh', width: '100%' }}>
       <DataGrid
       getRowId={(row) => row.internalId}
       rows={pdr}
@@ -196,8 +196,7 @@ export default function DataGridTable () {
       localeText={localeObj}
       components={{ Toolbar: GridToolbar }}
       componentsProps={{
-        toolbar: { showQuickFilter: true },
-        footer: {"data-testid": "footer"}
+        toolbar: { showQuickFilter: true }
       }}
       processRowUpdate={processRowUpdate}
       experimentalFeatures={{ newEditingApi: true }}/>
