+++ components/RadioButtons.js
@@ -1,27 +1,36 @@
 import { green, red, yellow } from '@mui/material/colors'
 import Radio from '@mui/material/Radio'
+import { withStyles } from '@mui/styles'
 import React from 'react'
 
-const GreenRadio = ({sx, ...props}) => <Radio sx={{
-  color: green[400],
-  '&.Mui-checked': {
-    color: green[600],
+const GreenRadio = withStyles({
+  root: {
+    color: green[400],
+    '&$checked': {
+      color: green[600]
+    }
   },
-...sx}} {...props}/>
+  checked: {}
+})((props) => <Radio color="default" {...props} />)
 
-const RedRadio = ({sx, ...props}) => <Radio sx={{
-  color: red[800],
-  '&.Mui-checked': {
-    color: red[600],
+const RedRadio = withStyles({
+  root: {
+    color: red[400],
+    '&$checked': {
+      color: red[600]
+    }
   },
-...sx}} {...props}/>
+  checked: {}
+})((props) => <Radio color="default" {...props} />)
 
-const YellowRadio = ({sx, ...props}) => <Radio sx={{
-  color: yellow[400],
-  '&.Mui-checked': {
-    color: yellow[600],
+const YellowRadio = withStyles({
+  root: {
+    color: yellow[400],
+    '&$checked': {
+      color: yellow[600]
+    }
   },
-...sx}} {...props}/>
+  checked: {}
+})((props) => <Radio color="default" {...props} />)
 
 export { RedRadio, YellowRadio, GreenRadio }
-
