+++ components/WeightDataGridTable.js
@@ -4,8 +4,8 @@ import { Button } from '@mui/material'
 import { DataGrid, esES, GridActionsCellItem, GridToolbarContainer } from '@mui/x-data-grid'
 import moment from 'moment'
 import { useCallback, useContext, useState } from 'react'
-import { WeightContext } from '../../context/WeightContext'
-import DeleteRowDialog from '../DeleteRowDialog'
+import { WeightContext } from '../context/WeightContext'
+import DeleteRowDialog from './DeleteRowDialog'
 
 function EditToolbar (props) {
   const { weight, setWeight } = props
@@ -90,7 +90,7 @@ export default function WeightDataGridTable (props) {
     { field: 'basura', headerName: 'Basura (lb)', editable: true, type: 'number', width: 100 }
   ]
   return (
-        <div style={{ height: '80vh', width: '100%' }} data-testid='weight-table'>
+        <div style={{ height: '80vh', width: '100%' }}>
             <DataGrid
             getRowId={(row) => moment(row.date).format('DD/MM/YYYY') }
             rows={weight}
@@ -100,8 +100,7 @@ export default function WeightDataGridTable (props) {
             experimentalFeatures={{ newEditingApi: true }}
             components={{ Toolbar: EditToolbar }}
             componentsProps={{
-              toolbar: { weight, setWeight },
-              footer: {"data-testid": "footer"}
+              toolbar: { weight, setWeight }
             }} />
             <DeleteRowDialog rowToDelete={rowToDelete} setRowToDelete={setRowToDelete} deleteRow={deleteRow} />
         </div>
