+++ package.json
@@ -6,39 +6,34 @@
     "build": "next build",
     "start": "next start -p 8080",
     "deploy": "npm run build && gcloud app deploy",
-    "lint": "next lint",
-    "test": "jest"
+    "lint": "next lint"
   },
   "dependencies": {
     "@emotion/react": "^11.7.1",
     "@emotion/styled": "^11.6.0",
     "@googlemaps/markerclusterer": "^2.0.10",
-    "@googlemaps/react-wrapper": "^1.1.35",
     "@material-table/core": "^0.2.35",
     "@material-table/exporters": "^1.0.12",
     "@mui/icons-material": "^5.3.1",
     "@mui/material": "^5.5.3",
+    "@mui/styles": "^5.5.3",
     "@mui/x-data-grid": "^5.12.3",
     "@mui/x-data-grid-generator": "^5.12.3",
     "devextreme": "^21.2.5",
     "eslint-plugin-unused-imports": "^2.0.0",
     "moment": "^2.29.4",
     "next": "12.0.10",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "recharts": "2.1.12"
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
+    "recharts": "^2.1.9"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
     "eslint": "^8.18.0",
     "eslint-config-next": "12.0.10",
     "eslint-config-standard": "^17.0.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-n": "^15.2.3",
     "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.30.1",
-    "jest": "^29.1.2",
-    "jest-environment-jsdom": "^29.1.2"
+    "eslint-plugin-react": "^7.30.1"
   }
 }
