+++ pages/downloadMap.js
@@ -1,9 +1,9 @@
 import Head from 'next/head'
+import DownloadMap from '../components/DownloadMap'
 import Layout from '../components/layout/Layout'
-import DownloadMap from '../components/map/DownloadMap'
 import styles from '../styles/Home.module.css'
 
-export default function DownloadMapPage () {
+export default function Map () {
   return (
     <div className={styles.container}>
       <Head>
