+++ pages/newPdr.js
@@ -1,9 +1,9 @@
 import Head from 'next/head'
 import Layout from '../components/layout/Layout'
-import NewPdr from '../components/newpdr/NewPdr'
+import NewPdr from '../components/NewPdr'
 import styles from '../styles/Home.module.css'
 
-export default function NewPdrPage () {
+export default function Map () {
   return (
     <div className={styles.container}>
       <Head>
