+++ pages/pasarPuntos.js
@@ -1,9 +1,9 @@
 import Head from 'next/head'
 import Layout from '../components/layout/Layout'
-import PasarPuntos from '../components/pasarpuntos/PasarPuntos'
+import PasarPuntos from '../components/PasarPuntos'
 import styles from '../styles/Home.module.css'
 
-export default function PasarPuntosPage () {
+export default function Map () {
   return (
     <div className={styles.container}>
       <Head>
